Implement CowUtils on Cow<str> itself to enable chaining
Chaining multiple calls to `CowUtils` methods will currently yield results tied
to temporary lifetimes because intermediate `Cow<'s, str>` will be borrowed as
`&'a str` via `&'a Cow<'s, str>` thereby loosing the original lifetime `'s`.

By implementing `CowUtils` for `Cow<str>` we can chain the source lifetime
through the calls at the cost of returning an owned `Cow::Owned(String)` instead
of a borrow `&'a str` if a previous operation modified the source value, i.e.
the calls in a chain will either all borrow or stop to completey at some point.

Another cost is that to handle short-lived references to owned values, the trait
itself must be extended to require a `Pattern` impl for any lifetime, not just
the given source lifetime. I don't think this is usually an issue, but it does
require a HRTB and is really only necessary for chaining.